split exercise tile helpers
