+++ lib/src/features/routines/data/repositories/workout_plan_repository_impl.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 import 'package:excel/excel.dart';
 import 'package:path_provider/path_provider.dart';
-
 import '../../domain/entities/exercise.dart';
 import '../../domain/entities/workout_plan.dart';
 import '../../domain/entities/plan_exercise_detail.dart';
@@ -9,53 +8,15 @@ import '../../domain/entities/workout_log_entry.dart';
 import '../../domain/entities/workout_session.dart';
 import '../../domain/repositories/workout_plan_repository.dart';
 import '../../../../data/schema/schemas.dart';
-
+part 'workout_plan_repository_utils.dart';
 class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
   List<Exercise>? _exerciseCache;
-  Future<File> _getOrCreateFile(String filename) async {
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/$filename');
-
-    if (!await file.exists()) {
-      final schema = kTableSchemas[filename]!;
-      final excel = Excel.createExcel();
-      final defaultSheet = excel.getDefaultSheet();
-      if (defaultSheet != null) {
-        excel.rename(defaultSheet, schema.sheetName);
-      }
-      excel[schema.sheetName]!.appendRow(
-        schema.headers.map<CellValue?>((e) => TextCellValue(e)).toList(),
-      );
-      final bytes = excel.save();
-      if (bytes != null) await file.writeAsBytes(bytes);
-    }
-    return file;
-  }
-
-  int _getLastId(Sheet sheet) {
-    for (var i = sheet.rows.length - 1; i >= 1; i--) {
-      final val = sheet.rows[i][0]?.value;
-      if (val != null) return int.tryParse(val.toString()) ?? 0;
-    }
-    return 0;
-  }
-  T? _cast<T>(Data? cell) {
-    final v = cell?.value;
-    if (v == null) return null;
-    if (v is T) return v as T;
-    if (T == int) return int.tryParse(v.toString()) as T?;
-    if (T == double) return double.tryParse(v.toString()) as T?;
-    if (T == String) return v.toString() as T;
-    return null;
-  }
-
   @override
   Future<List<WorkoutPlan>> getAllPlans() async {
     final file = await _getOrCreateFile('workout_plan.xlsx');
     final excel = Excel.decodeBytes(await file.readAsBytes());
     final sheet = excel[kTableSchemas['workout_plan.xlsx']!.sheetName];
     if (sheet == null) return [];
-
     return sheet.rows
         .skip(1)
         .where((row) => row.isNotEmpty)
@@ -67,13 +28,11 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
         })
         .toList();
   }
-
   @override
   Future<void> createWorkoutPlan(String name, String frequency) async {
     final file = await _getOrCreateFile('workout_plan.xlsx');
     final excel = Excel.decodeBytes(await file.readAsBytes());
     final sheet = excel[kTableSchemas['workout_plan.xlsx']!.sheetName]!;
-
     sheet.appendRow([
       IntCellValue(_getLastId(sheet) + 1),
       TextCellValue(name),
@@ -81,26 +40,21 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
     ]);
     await file.writeAsBytes(excel.save()!);
   }
-
   @override
   Future<List<Exercise>> getExercisesForPlan(int planId) async {
     final peFile = await _getOrCreateFile('plan_exercise.xlsx');
     final exFile = await _getOrCreateFile('exercise.xlsx');
-
     final peSheet = Excel.decodeBytes(await peFile.readAsBytes())[
         kTableSchemas['plan_exercise.xlsx']!.sheetName];
     final exSheet = Excel.decodeBytes(await exFile.readAsBytes())[
         kTableSchemas['exercise.xlsx']!.sheetName];
-
     if (peSheet == null || exSheet == null) return [];
-
     final exerciseIdsForPlan = peSheet.rows
         .skip(1)
         .where((row) => row.isNotEmpty && _cast<int>(row[0]) == planId)
         .map((row) => _cast<int>(row[1]))
         .whereType<int>()
         .toSet();
-
     return exSheet.rows
         .skip(1)
         .where((row) =>
@@ -118,16 +72,13 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
         )
         .toList();
   }
-
   @override
   Future<List<Exercise>> getAllExercises() async {
     if (_exerciseCache != null) return _exerciseCache!;
-
     final exFile = await _getOrCreateFile('exercise.xlsx');
     final exSheet =
         Excel.decodeBytes(await exFile.readAsBytes())[kTableSchemas['exercise.xlsx']!.sheetName];
     if (exSheet == null) return [];
-
     _exerciseCache = exSheet.rows
         .skip(1)
         .where((row) => row.isNotEmpty)
@@ -143,7 +94,6 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
         .toList();
     return _exerciseCache!;
   }
-
   @override
   Future<List<Exercise>> getSimilarExercises(int exerciseId) async {
     final all = await getAllExercises();
@@ -153,19 +103,15 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
         .where((e) => e.id != exerciseId && (e.category == base.category || e.mainMuscleGroup == base.mainMuscleGroup))
         .toList();
   }
-
   @override
   Future<List<PlanExerciseDetail>> getPlanExerciseDetails(int planId) async {
     final peFile = await _getOrCreateFile('plan_exercise.xlsx');
     final exFile = await _getOrCreateFile('exercise.xlsx');
-
     final peSheet = Excel.decodeBytes(await peFile.readAsBytes())[
         kTableSchemas['plan_exercise.xlsx']!.sheetName];
     final exSheet = Excel.decodeBytes(await exFile.readAsBytes())[
         kTableSchemas['exercise.xlsx']!.sheetName];
-
     if (peSheet == null || exSheet == null) return [];
-
     final mapIdName = {
       for (var r in exSheet.rows.skip(1))
         if (r.isNotEmpty) _cast<int>(r[0])!: r[1]?.value.toString() ?? '',
@@ -174,7 +120,6 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
       for (var r in exSheet.rows.skip(1))
         if (r.isNotEmpty) _cast<int>(r[0])!: r[2]?.value.toString() ?? '',
     };
-
     return peSheet.rows
         .skip(1)
         .where((r) =>
@@ -194,16 +139,12 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
         })
         .toList();
   }
-
-
   @override
   Future<void> saveWorkoutLogs(List<WorkoutLogEntry> logs) async {
     if (logs.isEmpty) return;
-
     final file = await _getOrCreateFile('workout_log.xlsx');
     final excel = Excel.decodeBytes(await file.readAsBytes());
     final sheet = excel[kTableSchemas['workout_log.xlsx']!.sheetName]!;
-
     var id = _getLastId(sheet) + 1;
     for (final l in logs) {
       sheet.appendRow([
@@ -219,13 +160,11 @@ class WorkoutPlanRepositoryImpl implements WorkoutPlanRepository {
     }
     await file.writeAsBytes(excel.save()!);
   }
-
   @override
   Future<void> saveWorkoutSession(WorkoutSession s) async {
     final file = await _getOrCreateFile('workout_session.xlsx');
     final excel = Excel.decodeBytes(await file.readAsBytes());
     final sheet = excel[kTableSchemas['workout_session.xlsx']!.sheetName]!;
-
     sheet.appendRow([
       IntCellValue(_getLastId(sheet) + 1),
       TextCellValue(s.date.toIso8601String().split('T').first),
